Delete reviews table migration file
Comments is now a self-referential association.

Change comment functionality of application. A 'review' or primary user reply of a song is now known as a comment. A comment can have many reply comments. There is no longer a distinction between a longer reply aka 'review' and a shorter reply aka 'comment'. When a user comments she can choose how she'd like to format the comment and the primary purpose of her comment. All comments on a song will be shown together. Comment replies will be nested.